feat(patch): enhance apply_patch function to include template name in error context
