+++ core/nemesis_merge/src/behavior/generate.rs
@@ -125,50 +125,3 @@ fn apply_and_gen_patched_hkx(owned_patches: &OwnedPatchMap, config: &Config) -> 
         hkx_errors: all_errors,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[allow(clippy::iter_on_single_items)]
-    const MODS: &[&str] = &[
-            // "../../dummy/Data/Nemesis_Engine/mod/aaaaa",
-            // "../../dummy/Data/Nemesis_Engine/mod/bcbi",
-            // "../../dummy/Data/Nemesis_Engine/mod/cbbi",
-            // "../../dummy/Data/Nemesis_Engine/mod/gender",
-            // "../../dummy/Data/Nemesis_Engine/mod/hmce",
-            // "../../dummy/Data/Nemesis_Engine/mod/momo",
-            // "../../dummy/Data/Nemesis_Engine/mod/na1w",
-            // "../../dummy/Data/Nemesis_Engine/mod/nemesis",
-            // "../../dummy/Data/Nemesis_Engine/mod/pscd",
-            // "../../dummy/Data/Nemesis_Engine/mod/rthf",
-            // "../../dummy/Data/Nemesis_Engine/mod/skice",
-            // "../../dummy/Data/Nemesis_Engine/mod/sscb",
-            // "../../dummy/Data/Nemesis_Engine/mod/tkuc",
-            // "../../dummy/Data/Nemesis_Engine/mod/tudm",
-            // "../../dummy/Data/Nemesis_Engine/mod/turn",
-            // "../../dummy/Data/Nemesis_Engine/mod/zcbe",
-            // "D:/GAME/ModOrganizer Skyrim SE/mods/FlinchingSSE やられモーションを追加(要OnHit/Nemesis_Engine/mod/flinch",
-            "D:/GAME/ModOrganizer Skyrim SE/mods/Crouch Sliding スプリント→しゃがみでスライディング/Nemesis_Engine/mod/slide",
-            "D:/GAME/ModOrganizer Skyrim SE/mods/Eating Animations And Sounds SE 歩行しながら食べるモーション/Nemesis_Engine/mod/eaas"
-    ];
-
-    #[tokio::test]
-    #[ignore = "unimplemented yet"]
-    #[cfg(feature = "tracing")]
-    async fn merge_test() {
-        let log_path = "../../dummy/merge_test.log";
-        crate::global_logger::global_logger(log_path, tracing::Level::TRACE).unwrap();
-
-        let _ = behavior_gen(
-            MODS.into_par_iter().map(|s| s.into()).collect(),
-            Config {
-                resource_dir: "../../resource/assets/templates".into(),
-                output_dir: "../../dummy/behavior_gen/output".into(),
-                status_report: Some(crate::config::new_color_status_reporter()),
-                hack_options: Some(crate::HackOptions::enable_all()),
-            },
-        )
-        .await;
-    }
-}
